fix(FileUploadList): cR comments
