+++ src/components/molecules/FileUploadList/FileUploadItem/FileUploadItem.scss
@@ -1,14 +1,12 @@
 .fileUploadItem {
     $this: &;
 
-    &__itemWrapper {
-        display: flex;
-        flex-direction: column;
-        row-gap: var(--guit-ref-spacing-3xsmall);
-        width: 100%;
-    }
+    display: flex;
+    flex-direction: column;
+    row-gap: var(--guit-ref-spacing-3xsmall);
+    width: 100%;
 
-    &__item {
+    &__content {
         display: flex;
         align-items: center;
         gap: var(--guit-ref-spacing-medium);
@@ -30,7 +28,7 @@
         }
     }
 
-    &__file {
+    &__iconWrapper {
         display: flex;
         align-items: center;
         justify-content: center;
