--- src/components/molecules/FileUploadList/FileUploadItem/types.ts
@@ -1,1 +1,0 @@
-export type FileType = "image" | "video" | "audio" | "file";
